requirements.txt
absl-py==1.1.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
dnspython==2.2.1
entrypoints==0.4
exchangelib==4.7.3
executing==0.8.3
flatbuffers==1.12
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.11.4
ipykernel==6.15.0
ipython==8.4.0
isodate==0.6.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
lxml==4.9.0
Markdown==3.3.7
matplotlib-inline==0.1.3
msal==1.18.0
nest-asyncio==1.5.5
ntlm-auth==1.5.0
numpy==1.23.0
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.0
requests==2.28.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.3.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.0
